Added some basic error handling

+++ run.py
@@ -3,6 +3,8 @@ import shutil
 import stat
 import subprocess
 import sys
+import time
+
 import pytest
 
 from git import Repo
@@ -18,7 +20,6 @@ def clone_repo(repo_url, target_dir):
 
 
 def make_cas():
-    print(os.getcwd())
     subprocess.run(['make', 'cas'])
 
 
@@ -28,6 +29,9 @@ if __name__ == "__main__":
     cas_repo_path = os.path.join("test_data", "cas")
     url = 'https://github.com/codenotary/cas.git'
 
+    if "CAS_API_KEY" not in os.environ:
+        sys.exit("CAS_API_KEY not set")
+
     if os.path.isdir(bin_dir):
         shutil.rmtree(bin_dir, onerror=remove_readonly)
         clone_repo(url, bin_dir)
@@ -35,7 +39,10 @@ if __name__ == "__main__":
         make_cas()
     else:
         clone_repo(url, bin_dir)
+        os.chdir(bin_dir)
+        make_cas()
+
+
 
     os.chdir(project_base_dir)
-    print(project_base_dir)
     sys.exit(pytest.main(['-s']))
